[Sonali] Add. should not remove product when id is invalid

+++ src/main/java/com/example/dailyneed_backened_repo/product/ProductService.java
@@ -96,7 +96,9 @@ public class ProductService {
         productRepository.save(product);
     }
 
-    public void removeProduct(Long id) {
+    public void removeProduct(Long id) throws ProductNotFoundException {
+        if(!checkIfProductExist(id))
+            throw new ProductNotFoundException();
         productRepository.deleteById(id);
     }
 }
